Moving area_data out of merc.h
